Don't attempt to regain exclusive lock in `backup`
When we've only got a shared lock, it means the cluster is in use, hence
immediately trying to get an exclusive lock is likely to block indefinitely.
However, except on the first run, we don't actually need an exclusive lock; a
shared lock will do nicely. We only need that exclusive lock if we're updating a
setting that needs a restart to make it active.

This change was extremely challenging because of Rust's borrowing and ownership
semantics. Appeasing the compiler was 95% of the work. Even then it prevented me
from explicitly releasing resources: it forbade me from calling `into_inner()`
on an `RwLock`; see the commentary in the diff for more information.